Reorganize csv_to_schema
